improve window xaml
