+++ src/WindowsApplication/MainWindow.xaml
@@ -1,24 +1,21 @@
-﻿<Window x:Class="WindowsApplication.MainWindow"
+﻿<utilities:InvisibleWindow x:Class="WindowsApplication.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:WindowsApplication"
 		xmlns:vm="clr-namespace:WindowsApplication.ViewModules"
+		xmlns:utilities="clr-namespace:WindowsApplication.Utilities"
 		xmlns:gif="https://github.com/XamlAnimatedGif/XamlAnimatedGif"
         mc:Ignorable="d"
         Title="MainWindow"
 		UseLayoutRounding="True"
 		Topmost="True"
         Focusable="False"
-        AllowsTransparency="True"
-        Background="Transparent"
         WindowStyle="None"
-        d:Visibility="Hidden"
-        ShowInTaskbar="False"
         SizeToContent="WidthAndHeight"    
         Loaded="Window_Loaded"
-        Name="Main"
+        x:Name="Main"
 		MouseMove="Icon_MouseMove"
 		MouseLeftButtonDown="Icon_MouseLeftButtonDown"
 		MouseLeftButtonUp="Icon_MouseLeftButtonUp"
@@ -33,10 +30,9 @@
 		   gif:AnimationBehavior.SourceUri="/Assets/ClickJusta_Animation.gif"
 		   gif:AnimationBehavior.AutoStart="False"
 		   Height="100"
-		   SnapsToDevicePixels="True"
 		   RenderOptions.BitmapScalingMode="HighQuality"
 		   RenderOptions.EdgeMode="Aliased"
 		   x:Name="img"
 		   Style="{StaticResource ClickableElement}"
 		   Stretch="Uniform" />
-</Window>
+</utilities:InvisibleWindow>
