+++ src/WindowsApplication/PresentationControl.xaml
@@ -8,7 +8,6 @@
              xmlns:local="clr-namespace:WindowsApplication"
              mc:Ignorable="d" 
              d:Background="White"
-             d:Visibility="Visible"
              Width="350"
              Height="150"
              Name="presentationCont"
